+++ projects/ngx-scrollbar/src/scrollbar/ng-scrollbar-thumb.ts
@@ -1,13 +1,15 @@
 import {
   Component,
   Inject,
-  Input,
-  ViewChild,
   OnInit,
   AfterViewInit,
   OnDestroy,
+  Input,
+  ViewChild,
+  HostBinding,
   NgZone,
   ElementRef,
+  ChangeDetectionStrategy,
   forwardRef
 } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
@@ -56,7 +58,7 @@ const axis: Axis = {
 
 @Component({
   selector: 'ng-scrollbar-thumb',
-  styleUrls: ['./ng-scrollbar-thumb.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
     <div #bar class="ng-scrollbar ng-scrollbar-{{orientation}} {{barClass}}" (mousedown)="onMouseDown($event)">
       <div #thumb class="ng-scrollbar-thumb {{thumbClass}}" [ngStyle]="scrollbarStyle | async"></div>
@@ -72,6 +74,9 @@ export class NgScrollbarThumb implements OnInit, AfterViewInit, OnDestroy {
 
   @ViewChild('bar') bar: ElementRef;
   @ViewChild('thumb') thumb: ElementRef;
+  @HostBinding('class.ng-scrollbar-visible') get visibility(): boolean {
+    return !!this._scrollMax;
+  }
 
   private _minThumbSize = 20;
   private _naturalThumbSize = 0;